Refactor to modularize graph vis.
