+++ tensortapestry-loom/src/main/java/org/tensortapestry/loom/graph/LoomNodeWrapper.java
@@ -3,6 +3,7 @@ package org.tensortapestry.loom.graph;
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import java.util.UUID;
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import org.tensortapestry.common.collections.Wrapper;
 
 /**
@@ -10,6 +11,16 @@ import org.tensortapestry.common.collections.Wrapper;
  */
 public interface LoomNodeWrapper<WrapperT extends LoomNodeWrapper<WrapperT>>
   extends Wrapper<LoomNode> {
+  @Nonnull
+  default LoomGraph assertGraph() {
+    return unwrap().assertGraph();
+  }
+
+  @Nonnull
+  default LoomEnvironment assertEnvironment() {
+    return unwrap().assertEnvironment();
+  }
+
   /**
    * Get the ID of the node.
    *
@@ -18,6 +29,15 @@ public interface LoomNodeWrapper<WrapperT extends LoomNodeWrapper<WrapperT>>
   @Nonnull
   default UUID getId() {
     return unwrap().getId();
+  }
+
+  /**
+   * Get the label of the node.
+   *
+   * @return the label.
+   */
+  @Nullable default String getLabel() {
+    return unwrap().getLabel();
   }
 
   @Nonnull
