+++ tensortapestry-loom/src/main/java/org/tensortapestry/loom/graph/dialects/tensorops/OperationExpressionDialect.java
@@ -1,5 +1,8 @@
 package org.tensortapestry.loom.graph.dialects.tensorops;
 
+import guru.nidi.graphviz.engine.Format;
+import guru.nidi.graphviz.engine.Graphviz;
+import java.awt.*;
 import lombok.experimental.UtilityClass;
 import org.tensortapestry.loom.graph.CommonEnvironments;
 import org.tensortapestry.loom.graph.LoomEnvironment;
@@ -9,6 +12,8 @@ import org.tensortapestry.loom.graph.dialects.common.TypeRestrictionConstraint;
 import org.tensortapestry.loom.graph.dialects.tensorops.constraints.NoTensorOperationCyclesConstraint;
 import org.tensortapestry.loom.graph.dialects.tensorops.constraints.OperationIPFSignatureAgreementConstraint;
 import org.tensortapestry.loom.graph.dialects.tensorops.constraints.TensorOperationAgreementConstraint;
+import org.tensortapestry.loom.graph.export.graphviz.LoomGraphvizExporter;
+import org.tensortapestry.loom.graph.export.graphviz.OperationExpressionLoomGraphvizExporter;
 
 @UtilityClass
 public class OperationExpressionDialect {
@@ -37,5 +42,21 @@ public class OperationExpressionDialect {
 
   public LoomGraph newGraph() {
     return ENVIRONMENT.newGraph();
+  }
+
+  public LoomGraphvizExporter getGraphvizExporter() {
+    return OperationExpressionLoomGraphvizExporter.builder().build();
+  }
+
+  public LoomGraphvizExporter.ExportContext export(LoomGraph graph) {
+    return getGraphvizExporter().export(graph);
+  }
+
+  public Graphviz toGraphviz(LoomGraph graph) {
+    return export(graph).toGraphvizWrapper();
+  }
+
+  public Image toImage(LoomGraph graph) {
+    return toGraphviz(graph).render(Format.PNG).toImage();
   }
 }
