+++ tensortapestry-loom/src/main/java/org/tensortapestry/loom/graph/export/graphviz/UuidAliasEnv.java
@@ -1,9 +1,9 @@
 package org.tensortapestry.loom.graph.export.graphviz;
 
 import java.util.*;
-import java.util.function.Function;
+import org.tensortapestry.zspace.indexing.IndexingFns;
 
-public class UuidAliasProvider implements Function<UUID, String> {
+public final class UuidAliasEnv {
 
   private static final List<String> VOCAB1 = List.of(
     "red",
@@ -32,31 +32,27 @@ public class UuidAliasProvider implements Function<UUID, String> {
 
   private final Map<UUID, String> aliasMap;
 
-  public UuidAliasProvider() {
+  public UuidAliasEnv() {
     this(new HashMap<>());
   }
 
-  public UuidAliasProvider(Map<UUID, String> aliasMap) {
+  public UuidAliasEnv(Map<UUID, String> aliasMap) {
     this.aliasMap = new HashMap<>(aliasMap);
   }
 
-  private static int abs(int x) {
-    if (x < 0) {
-      return -x;
-    }
-    return x;
+  public boolean hasIdAlias(String alias) {
+    return aliasMap.containsValue(alias);
   }
 
-  @Override
-  public synchronized String apply(UUID uuid) {
+  public synchronized String getIdAlias(UUID uuid) {
     if (!aliasMap.containsKey(uuid)) {
-      int h = abs(uuid.hashCode());
+      int h = IndexingFns.throwOnMinAbs(uuid.hashCode());
       var base =
         VOCAB1.get(h % VOCAB1.size()) + ":" + VOCAB2.get((h / VOCAB1.size()) % VOCAB2.size());
 
       for (int i = 1;; i++) {
         var alias = base + ":" + i;
-        if (!aliasMap.containsValue(alias)) {
+        if (!hasIdAlias(alias)) {
           aliasMap.put(uuid, alias);
           return alias;
         }
